requirements.txt
# Scientific Computing Stack
numpy>=1.19.0
scipy>=1.5.0
pandas>=1.1.0
matplotlib>=3.3.0

# Gravitational Wave Analysis
gwpy>=2.0.0
python-ligo-lw>=1.7.0

# Pipeline & Tools (Documentation purpose for Portfolio)
hveto>=1.0.1
beautifulsoup4>=4.9.0
torch>=1.7.0
torchvision>=0.8.0
tensorflow>=2.5.0
Pillow>=8.0.0
